pyproject.toml
[tool.poetry]
name = "multimodal-memory"
version = "0.1.0"
description = ""
authors = ["Christos <chziakas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^0.28.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
